move files to separate folder
