+++ clients/flutter_client/lib/src/widgets/devices_view.dart
