+++ clients/flutter_client/lib/src/widgets/home_default_view.dart
